### 2023-03-23 #5279 @erzk
Hi @oliviaguest

Thanks for reviewing the paper. The package code and paper are stored in separate repos.

The software is here https://github.com/ropensci/PostcodesioR
and the paper is here https://github.com/erzk/PostcodesioR_joss/tree/main/PostcodesioR
