### 2023-03-26 #5279 @oliviaguest
> Hi @oliviaguest
>
> Thanks for reviewing the paper. The package code and paper are stored in separate repos.
>
> The software is here https://github.com/ropensci/PostcodesioR and the paper is here https://github.com/erzk/PostcodesioR_joss/tree/main/PostcodesioR

Oh, I have no idea if @editorialbot can deal with this. I'll find out!
