### 2023-03-27 #5279 @oliviaguest
@erzk I actually think the bot cannot cope with this ([please read the docs if need be](https://joss.readthedocs.io/en/latest/editorial_bot.html#compiling-papers-from-a-specific-branch)) — is there any chance you can use a branch? If the problem is you prefer not to have such a branch, you may delete it after publication. 😊

**Also, please remind yourself of [JOSS's submission requirements](https://joss.readthedocs.io/en/latest/submitting.html?highlight=paper%20branch#submission-requirements), which actually stipulate this (branch requirements) quite clearly:**

> Your paper (paper.md and BibTeX files, plus any figures) must be hosted in a Git-based repository together with your software (although they may be in a short-lived branch which is never merged with the default).

I am happy to answer any further questions, _but please first make sure @editorialbot can compile your paper. ✨_
