+++ paper.bib
@@ -1,0 +1,17 @@
+@online{ONS:2023,
+  author = {{Office for National Statistics}},
+  title = {{Our changing geography}},
+  year = 2023,
+  url = {https://www.ons.gov.uk/methodology/geography/ukgeographies/administrativegeography/ourchanginggeography},
+  urldate = {2023-03-19}
+}
+
+@book{Lovelace:2019,
+  	Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+  	Adsurl = {http://adsabs.harvard.edu/abs/2008gady.book.....B},
+  	Author = {{Lovelace}, R. and {Nowosad}, J. and {Muenchow}, J.},
+  	Booktitle = {Geocomputation with R. The R Series.},
+  	Publisher = {Princeton University Press},
+  	Title = {{CRC Press}},
+  	Year = 2019
+}
