+++ .github/workflows/draft-pdf.yml
@@ -12,12 +12,12 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: PostcodesioR.md
+          paper-path: paper.md
       - name: Upload
         uses: actions/upload-artifact@v1
         with:
-          name: PostcodesioR
+          name: paper
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: PostcodesioR.pdf
+          path: paper.pdf
