### 2023-04-03 #5334 @editorialbot
[REVIEW]: PostcodesioR: An R package for UK geocoding
**Submitting author:** @erzk (<a href="http://orcid.org/0000-0001-8724-462X">Eryk Walczak</a>)
**Repository:** https://github.com/ropensci/PostcodesioR
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.3.1
