+++ DESCRIPTION
@@ -17,7 +17,7 @@ Description: Free UK geocoding using data from Office for National Statistics.
     It is using several functions to get information about post codes, outward
     codes, reverse geocoding, nearest post codes/outward codes, validation, or
     randomly generate a post code. API wrapper around <https://postcodes.io>.
-License: GPL-3
+License: GPL (>= 3)
 URL: https://docs.ropensci.org/PostcodesioR/
 Depends:
     R (>= 3.1)
