### 2023-04-08 #5334 @erzk
@oliviaguest

- [x] GPL3 license file added
- [x] JOSS and Zenodo titles are matching
- [x]  JOSS and Zenodo author is matching. Zenodo must have requested the list of all minor contributors and reviewers so it showed additional authors.

The paper looks finished. Are any additional changes that are required?
Thanks.
