+++ paper.bib
@@ -10,8 +10,8 @@
   	Adsnote = {Provided by the SAO/NASA Astrophysics Data System},
   	Adsurl = {http://adsabs.harvard.edu/abs/2008gady.book.....B},
   	Author = {{Lovelace}, R. and {Nowosad}, J. and {Muenchow}, J.},
-  	Booktitle = {Geocomputation with R. The R Series.},
-  	Publisher = {Princeton University Press},
-  	Title = {{CRC Press}},
+  	Booktitle = {Geocomputation with R. The R Series},
+  	Publisher = {{CRC Press}},
+  	Title = {{Geocomputation with R. The R Series}},
   	Year = 2019
 }
