### 2023-04-14 #5334 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
